Unregister the enchantment if previously registered
When the server is reloaded the Enchantments are serialized and unserialized.  When it is unserialized it is unserialized as an Enchantment and therefore no longer an instance of CustomEnchantment and any custom listeners will not function.  If the enchantment is already registered then it will not be overwritten and will remain an Enchantment.  Unregistering ensures that the stored instance is an instance of CustomEnchantment.